+++ dapp/src/components/Banner.js
@@ -1,0 +1,198 @@
+import React from 'react'
+import classnames from 'classnames'
+import Link from 'next/link'
+import { useRouter } from 'next/router'
+import { useStoreState } from 'pullstate'
+import { fbt } from 'fbt-runtime'
+import moment from 'moment'
+import StakeStore from 'stores/StakeStore'
+import { adjustLinkHref } from 'utils/utils'
+
+const Banner = ({ dapp }) => {
+  const burn = moment('2022-10-10T00:00:00.000Z')
+  const days = burn.diff(moment(), 'days')
+  const burnDays = days === 0 ? 1 : days
+
+  const { pathname } = useRouter()
+  const burnPage = pathname === '/burn'
+  const stakePage = pathname === '/earn'
+  const stakes = useStoreState(StakeStore, (s) => s)
+  const showStakingBanner =
+    dapp && (stakes.stakes || []).length !== 0 && !stakePage
+
+  const notice = showStakingBanner || burnDays >= 0
+
+  const StakingBanner = () => {
+    return (
+      <>
+        <div className="d-flex flex-column mt-0 justify-content-center px-4 px-md-0 text-md-left">
+          <div className="title-text">
+            {fbt(
+              'Changes are coming to OGN staking.',
+              'Changes are coming to OGN staking.'
+            )}
+          </div>
+          <div className="text">
+            {fbt(
+              'Your existing stakes will not be impacted. Claim your OGN at the end of your staking period.',
+              'Your existing stakes will not be impacted. Claim your OGN at the end of your staking period.'
+            )}
+          </div>
+        </div>
+        <div className="btn btn-dark mt-2 ml-md-auto">
+          <Link href={adjustLinkHref('/earn')}>Legacy staking</Link>
+        </div>
+        <style jsx>{`
+          .btn {
+            font-size: 12px;
+            height: auto;
+            padding: 5px 20px;
+            background-color: white;
+            color: black;
+          }
+
+          .title-text {
+            font-size: 18px;
+            font-weight: bold;
+            line-height: 1.75;
+            color: white;
+          }
+
+          .text {
+            opacity: 0.8;
+            color: white;
+            line-height: normal;
+            font-size: 14px;
+            max-width: 1000px;
+          }
+        `}</style>
+      </>
+    )
+  }
+  const ClaimBanner = () => {
+    const AIRDROP_URL = 'https://governance.ousd.com/claim'
+
+    return (
+      <>
+        {fbt('OGV airdrop is live!', 'Airdrop notice')}
+        <a
+          href={AIRDROP_URL}
+          target="_blank"
+          rel="noopener noreferrer"
+          className="btn btn-dark mt-3 mt-md-0 ml-md-auto"
+        >
+          Check eligibility
+        </a>
+        <style jsx>{`
+          .btn {
+            font-size: 12px;
+            height: auto;
+            padding: 5px 20px;
+            background-color: white;
+            color: black;
+          }
+        `}</style>
+      </>
+    )
+  }
+  const BurnBanner = () => {
+    return (
+      <>
+        {fbt(
+          'Only ' +
+            fbt.param('burn-days', burnDays) +
+            ' days left to claim your OGV before the burn',
+          'Burn notice'
+        )}
+        <Link href={adjustLinkHref('/burn')}>
+          <a className="btn btn-dark gradient2 mt-3 mt-md-0 ml-md-auto">
+            OGV Burn
+          </a>
+        </Link>
+        <style jsx>{`
+          .btn {
+            font-size: 12px;
+            height: auto;
+            padding: 5px 20px;
+            color: black;
+          }
+        `}</style>
+      </>
+    )
+  }
+  return (
+    <>
+      {notice && (
+        <div
+          className={classnames(
+            `notice ${showStakingBanner ? 'staking pt-2' : 'pt-3'} ${
+              burnPage ? 'burn' : ''
+            } text-white text-center pb-3`,
+            {
+              dapp,
+            }
+          )}
+        >
+          <div className="container d-flex flex-column flex-md-row align-items-center">
+            {showStakingBanner ? (
+              <StakingBanner />
+            ) : burnPage ? (
+              <ClaimBanner />
+            ) : (
+              <BurnBanner />
+            )}
+          </div>
+        </div>
+      )}
+      <style jsx>{`
+        .notice {
+          background-color: black;
+          margin-bottom: 0px;
+        }
+
+        .notice.burn {
+          background: linear-gradient(90deg, #8c66fc -28.99%, #0274f1 144.97%);
+        }
+
+        .notice.staking {
+          background-color: #1a82ff;
+        }
+
+        .notice.dapp {
+          margin-bottom: 0px;
+        }
+
+        .notice .btn {
+          font-size: 12px;
+          height: auto;
+          padding: 5px 20px;
+          background-color: white;
+          color: black;
+        }
+
+        .container {
+          max-width: 940px;
+          padding-left: 0px;
+          padding-right: 0px;
+        }
+
+        .title-text {
+          font-size: 18px;
+          font-weight: bold;
+          line-height: 1.75;
+          color: white;
+        }
+
+        .text {
+          opacity: 0.8;
+          color: white;
+          line-height: normal;
+          font-size: 14px;
+          max-width: 1000px;
+        }
+      `}</style>
+    </>
+  )
+}
+
+export default Banner
